refactor: Introduce Ticket interface
Lotto 클래스가 Ticket 인터페이스를 구현하도록 리팩토링했습니다.
구조 개선과 확장성을 고려하여 추상화를 추가했습니다.
