+++ src/main/java/ticket/Ticket.java
@@ -1,0 +1,5 @@
+package ticket;
+
+public interface Ticket {
+    String getTicketInfo();
+}
